chore: update dependencies

+++ crates/lipl-display-freya/src/constant.rs
@@ -2,8 +2,8 @@ pub const APPLICATION_TITLE: &str = "Lipl Display";
 pub const APPLICATION_WIDTH: f64 = 600.0;
 pub const APPLICATION_HEIGHT: f64 = 300.0;
 pub const WAIT_MESSAGE: &str = "Even geduld a.u.b. ...";
-pub const MINIMUM_FONT_SIZE: usize = 4;
-pub const FONT_SIZE_INCREMENT: usize = 2;
+pub const MINIMUM_FONT_SIZE: i32 = 4;
+pub const FONT_SIZE_INCREMENT: i32 = 2;
 
 #[allow(dead_code)]
 pub const PATH: &str = "/home/paul/Code/dart/lipl_display/lipl-gatt-input.txt";
